Temp/review base (#1305)
* reviewed

* Baselines for https://github.com/linq2db/linq2db/pull/4676 (#1309)

* questionable

* [Linux / PostgreSQL 14] baselines

* [Linux / PostgreSQL 13] baselines

* [Linux / PostgreSQL 15] baselines

* [Linux / PostgreSQL 16] baselines

* [Linux / PostgreSQL 17] baselines

* [Linux / PostgreSQL 13] baselines

* [Linux / Oracle 23c] baselines

* [Linux / PostgreSQL 14] baselines

* [Linux / PostgreSQL 15] baselines

* [Linux / PostgreSQL 16] baselines

* [Linux / PostgreSQL 17] baselines

* [Linux / SQLite (both providers)] baselines

* [Linux / Oracle 18c] baselines

* [Linux / Oracle 23c] baselines

* [Linux / SQLite (both providers)] baselines

---------


---------

+++ Informix.DB2.LinqService/Tests/UserTests/Issue1284Tests/Tests.UserTests.Issue1284Tests.TestCteInvalidMappingUnion(Informix.DB2.LinqService).sql
@@ -22,7 +22,7 @@ AS
 	FROM
 		(
 			SELECT
-				1 as rn,
+				1::Int as rn,
 				x.FirstName,
 				x.PersonID as ID,
 				x.LastName,
@@ -32,7 +32,7 @@ AS
 				Person x
 			UNION ALL
 			SELECT
-				2 as rn,
+				2::Int as rn,
 				person_1.FirstName,
 				person_1.PersonID as ID,
 				person_1.LastName,
@@ -67,7 +67,7 @@ SELECT FIRST 1
 FROM
 	(
 		SELECT
-			1 as rn,
+			1::Int as rn,
 			x.FirstName,
 			x.PersonID as ID,
 			x.LastName,
@@ -77,7 +77,7 @@ FROM
 			Person x
 		UNION ALL
 		SELECT
-			2 as rn,
+			2::Int as rn,
 			person_1.FirstName,
 			person_1.PersonID as ID,
 			person_1.LastName,
